src/worker/requirements.txt
yfinance
pandas
sqlalchemy
psycopg2-binary
python-dotenv
schedule